refactor: move tests to src
