+++ plugin-server/src/main/ingestion-queues/session-recording-v2/teams/team-filter.test.ts
@@ -1,7 +1,7 @@
-import { ParsedMessageData } from '../../../../../src/main/ingestion-queues/session-recording-v2/kafka/types'
-import { TeamFilter } from '../../../../../src/main/ingestion-queues/session-recording-v2/teams/team-filter'
-import { TeamService } from '../../../../../src/main/ingestion-queues/session-recording-v2/teams/team-service'
-import { Team } from '../../../../../src/main/ingestion-queues/session-recording-v2/teams/types'
+import { ParsedMessageData } from '../kafka/types'
+import { TeamFilter } from './team-filter'
+import { TeamService } from './team-service'
+import { Team } from './types'
 
 jest.mock('../../../../../src/main/ingestion-queues/session-recording-v2/teams/team-service')
 
@@ -34,7 +34,7 @@ describe('TeamFilter', () => {
         jest.clearAllMocks()
         mockTeamService = {
             getTeamByToken: jest.fn(),
-        } as jest.Mocked<TeamService>
+        } as unknown as jest.Mocked<TeamService>
         teamFilter = new TeamFilter(mockTeamService)
     })
